+++ std-bits/database/src/main/java/org/enso/database/postgres/PostgresConnectionDetailsSPI.java
@@ -1,4 +1,6 @@
-package org.enso.database;
+package org.enso.database.postgres;
+
+import org.enso.database.DatabaseConnectionDetailsSPI;
 
 @org.openide.util.lookup.ServiceProvider(service = DatabaseConnectionDetailsSPI.class)
 public class PostgresConnectionDetailsSPI extends DatabaseConnectionDetailsSPI {
