+++ std-bits/database/src/main/java/org/enso/database/sqlite/SQLiteFormatSPI.java
@@ -1,4 +1,4 @@
-package org.enso.database;
+package org.enso.database.sqlite;
 
 import org.enso.base.file_format.FileFormatSPI;
